+++ email/emailjs.js
@@ -1,4 +1,11 @@
 'use strict';
+/**
+ * Kado - Web Application System
+ * Copyright © 2015-2019 Bryan Tong, NULLIVEX LLC. All rights reserved.
+ * Kado <support@kado.org>
+ *
+ * This file is part of Kado and bound to the MIT license distributed within.
+ */
 const email = require('emailjs')