+++ kado_modules/doc/admin/index.js
@@ -1,23 +1,10 @@
 'use strict';
 /**
- * Kado - Module system for Enterprise Grade applications.
- * Copyright © 2015-2019 NULLIVEX LLC. All rights reserved.
+ * Kado - Web Application System
+ * Copyright © 2015-2019 Bryan Tong, NULLIVEX LLC. All rights reserved.
  * Kado <support@kado.org>
  *
- * This file is part of Kado.
- *
- * Kado is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Kado is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Kado.  If not, see <https://www.gnu.org/licenses/>.
+ * This file is part of Kado and bound to the MIT license distributed within.
  */
 const crypto = require('crypto')
 
@@ -49,6 +36,8 @@ exports.version = require('./version')
  */
 exports.list = (req,res) => {
   if(!req.query.length){
+    res.locals._asset.addScript('/dist/dataTables.js')
+    res.locals._asset.addScript('/js/dataTableList.js')
     res.render('doc/list',{
       _pageTitle: K._l.doc.doc + ' ' + K._l.list})
   } else {
@@ -89,6 +78,8 @@ exports.create = (req,res) => {
  * @param {object} res
  */
 exports.edit = (req,res) => {
+  res.locals._asset.addScript('/dist/tuiEditor.js')
+  res.locals._asset.addScript('/js/loadTuiEditor.js')
   Doc.findByPk(req.query.id,res.Q)
     .then((result) => {
       if(!result) throw new Error(K._l.doc.entry_not_found)