+++ lang/spa.js
@@ -1,23 +1,10 @@
 'use strict';
 /**
- * Kado - Module system for Enterprise Grade applications.
- * Copyright © 2015-2019 NULLIVEX LLC. All rights reserved.
+ * Kado - Web Application System
+ * Copyright © 2015-2019 Bryan Tong, NULLIVEX LLC. All rights reserved.
  * Kado <support@kado.org>
  *
- * This file is part of Kado.
- *
- * Kado is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Kado is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with Kado.  If not, see <https://www.gnu.org/licenses/>.
+ * This file is part of Kado and bound to the MIT license distributed within.
  */
 
 //Kado english language pack
@@ -32,7 +19,7 @@ module.exports = {
   '_pack_flag': 'es',
   'admin_panel': 'Panel',
   'kado': 'Kado',
-  'kado_description': 'Sistema de desarrollo para aplicaciones web..',
+  'kado_description': 'Sistema de aplicación web.',
   'version': 'Versión',
   'welcome_to': 'Bienvenido a',
   'an_error_occurred': 'Se ha producido un error :(',